backend/requirements.txt
Flask==2.0.3
Flask-CORS==3.0.10
python-dotenv==0.19.2
openai-whisper
